Added images.rs to app_resources.rs
